fix compiler warning

+++ src/scip/lpi_qso.c
@@ -330,7 +330,7 @@ SCIP_RETCODE convertSides(
 const char* SCIPlpiGetSolverName(void)
 {
    char* vname = QSversion();
-   int vnamelen;
+   size_t vnamelen;
    vnamelen = strlen(vname);
    memcpy(__qsstr, vname, MIN(sizeof(__qsstr), vnamelen+1));
    __qsstr[sizeof(__qsstr)-1] = '\0';
